requirements.txt
gym==0.17.2
graphviz==0.19.1
numpy==1.19.0
pygame==2.5.2
pythomata==0.3.2
pyparsing==3.1.1
scikit-learn==0.23.1
scipy==1.5.1
tensorboard==1.14.0
tensorboardX==2.1
torch==1.5.1
clize==5.0.2
ltlf2dfa==1.0.2
matplotlib==3.3.4
PyYAML==6.0.1
tqdm==4.64.1
